+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 * 1.0.0-beta8 (@release_date@)
 
+ * changed `puli bind` command to generate `ClassBinding` instances when passing
+   a path that contains backslashes (`\`) or when passing the `--class` option
  * added `--class` option to `puli bind --update` command
  * added `puli upgrade` command